Model-database resolved bugs for mapping

+++ src/main/java/com/wediscussmovies/project/exception/MovieIdNotFoundException.java
@@ -1,7 +1,7 @@
 package com.wediscussmovies.project.exception;
 
 public class MovieIdNotFoundException extends RuntimeException{
-    public MovieIdNotFoundException(Long id){
+    public MovieIdNotFoundException(Integer id){
         super("Movie with that id=" + id + " does not exist!");
     }
 }
